Use correct pairing commitment
This fixes a bug in the server where the closure used to generate a fresh
compatible mix was closing over a loop iterator.  This would cause reruns to
create coinjoins with invalid amounts.  These issues were being correctly
detected by dcrwallet, which would abort the mix for not finding a correct mixed
message of the correct value.