Double client read timeouts
An issue was observed where a wallet successfully signed a mix and
responded to the server, but failed to read the resulting transaction
with all combined signatures.  It's suspected that this wallet was
faster to sign and respond to the mix, causing it to set an earlier
deadline than the server was able to respond within.  Doubling the
recv timeout should help.