Only generate fresh SR/DC messages when needed.
If a run fails due to a peer dropping out after the key exchange,
secrets do not need to be revealed for the peer to be excluded from
future runs.  While these values remain secret, continue to use them
for future runs.  This can result in less new addresses created by a
wallet when a non-misbehaving peer must be excluded from a mix due to
network issues.