Reconnect dcrd RPC client after client failure
If the RPC client is put into an error state for any reason (e.g.
network loss), there must be a mechanism to reconnect without
destroying the server.  This change creates and caches a RPC client
from the main package, continuing to use it so long as it hasn't
failed.  After failure, a new client is created each time a new mix
object must be obtained by the server.

While here, update to latest wsrpc, which adds support for and enables
websocket ping/pong by default.