Remove license badge
The license is already mentioned at the bottom of the readme.

Furthermore, this badge linked to a site that has absolutely nothing
to do with the ISC license, making me unsure why it was being linked
at all to begin with.

I have nothing good nor bad to say regarding the content at this link,
but it's not relevant to our project in any way, and theoretically it
is possible for it to be promoting ideas we do not subscribe to.  Just
read the license itself instead (it's not long or complicated).