Copy pairing queue's client slice for session
Sharing the slice of clients between the both the pairing queue and
the session was not safe.  It permitted a data race (the same
structure was being protected by two different mutexes).  In addition,
if a peer was removed from the pairing queue, removing the client
would change the order of all peers in the session and cause the
incorrect peers to be dropped out of mixes.