Blame and remove all peers with KE errors
When both KE timeout occurred for one peer, and a second peer double
spent the unmixed inputs previously shown to the server, only the
double spending peer was removed from the mix.  Keeping the timed-out
client in later runs caused server hangs when attempting to send rerun
messages, because the client's out message channel was not being read
anymore.