Add h2 to NextProtos
Go 1.17's crypto/tls package introduced strict ALPN checks.  When the
NextProtos field of a tls.Config is set by both client and server, the
client must provide a matching protocol (with a http2 -> http/1.1
downgrade exception).

The missing "h2" protocol was preventing the TLS web server in
csppserver working properly with browsers that provide their supported
protocols (h2 and http/1.1, in the case of Chromium).