Implement PQ key exchange
This implements a protocol change adding an additional round of
communication to share Streamlined NTRU Prime 4591^761 ciphertexts
between all peers.  This key exchange occurs in addition to an x25519
key exchange as previously, such that if a weakness in sntrup4591761
is discovered, the security won't reduce to anything worse than just
x25519.

Due to the large size of sntrup4591761 public keys and ciphertexts,
only a single key is exchanged between any two peers, regardless of
how many messages are being mixed.  The resulting shared key is used
to seed a shared ChaCha20 CSPRNG which derives shared keys for the SR
and DC mixes.  This differs from the previous key exchange where
x25519 keys were exchanged for every message in the DC-net.

The module's major version is bumped to v2 to represent the breaking
change.  There is no other API change for consumers, but v2 clients
are not compatible with v1 servers and vice-versa.