Add server-side check for insanely high fees.
Goal is to prevent a situation like decred/dcrwallet#2000 from
happening again even if users are running the buggy client code.

While here, the fee calculation is fixed to consider the cost of input
scripts to redeem P2PKH outputs.  Even if this is not the case,
minimum fee requirements in dcrd use the same assumption.