Sign big integers in a machine-independent way
This fixes a protocol bug which caused signature validation failures
if the client and server used different sizes of the uint type, such
as a 32-bit client connecting to a 64-bit server.

While here, also commit to the signedness of the big.Int.

This fix is a protocol breaking change.