Add csppsolver command and solverrpc package
The solverrpc package provides a drop-in replacement for solver.Roots,
so long as the csppsolver executable can be looked up in PATH.  It
runs csppsolver as a background process, performing RPC with it over
standard input/output.

This addition will enable pure-Go processes to run a solver built with
cgo.